stepping into SQL
I have 2 projects in this, one is very simple however I created the data that built into it understanding categories and how SQL operates. The second project was a lot more in-depth I worked with the top 100 movies and their statistics breaking them down in ways that built up on each other, challenging myself with functions I wasn't as confident with.