3 table join functions
for this project, I researched artists that came on shuffle, the artist and their bday, the songs that were in the playlist followed by the album, and where the artist is from. With 3 tables I messed with the join function to produce tables that incorporated two or even all three of the tables.